Move & alias connectivity-map command
