+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/backend/checkout/configmap.yaml
