+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/backend/checkout/deployment.yaml
