+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/frontend/asset-cache/route.yaml
