+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/frontend/webapp/configmap.yaml
