+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/payments/gateway/deployment.yaml
