+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/payments/mastercard-processor/deployment.yaml
