+++ roxctl/netpol/connectivity/map/testdata/acs-security-demos/payments/visa-processor/deployment.yaml
