+++ roxctl/netpol/connectivity/map/testdata/dirty/frontend.yaml
