+++ roxctl/netpol/connectivity/map/testdata/empty-yamls/empty.yaml
