+++ roxctl/netpol/connectivity/map/testdata/empty-yamls/empty2.yaml
