+++ roxctl/netpol/connectivity/map/testdata/frontend-security/asset-cache-deployment.yaml
