+++ roxctl/netpol/connectivity/map/testdata/frontend-security/asset-cache-route.yaml
