+++ roxctl/netpol/connectivity/map/testdata/frontend-security/frontend-netpols.yaml
