+++ roxctl/netpol/connectivity/map/testdata/frontend-security/webapp-deployment.yaml
