+++ roxctl/netpol/connectivity/map/testdata/minimal-with-invalid-doc/resources.yaml
