+++ roxctl/netpol/connectivity/map/testdata/minimal/frontend.yaml
