+++ roxctl/netpol/connectivity/map/testdata/minimal/netpols.yaml
