+++ roxctl/netpol/connectivity/map/testdata/minimal/output.txt
